requirements.txt
bootstrap4==0.1.0
Django==3.1
django-crispy-forms==1.9.2
Pillow==7.2.0
psycopg2==2.8.5
requests==2.22.0


